Move used by both deployment_handler and statefulset_handler, out of deployment_handler.go file
